+++ R/location_series.R
@@ -19,7 +19,7 @@
 #' @param bio_variables vector of names of variables to be extracted.
 #' @param dataset string defining the dataset to use. If set to "custom",
 #' then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included in
 #' this file.
 #'