+++ R/region_slice.R
@@ -13,10 +13,10 @@
 #' @param bio_variables vector of names of variables to be extracted
 #' @param dataset string defining the dataset to use. If set to "custom",
 #' then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included
 #' in this file.
-#' @param ext an extent, coded as numberic vector (length=4; 
+#' @param ext an extent, coded as numeric vector (length=4; 
 #' order= xmin, xmax, ymin, ymax) or a \code{terra::SpatExtent} object. If NULL,
 #' the full extent of the reconstruction is given.
 #' @param crop a polygon used to crop the reconstructions (e.g. the outline